Put the ending calculation in a different file.
