+++ server.js
@@ -7,6 +7,7 @@ const bodyParser = require('body-parser');
 const port = process.env.PORT || 3000;
 const countVotes = require('./lib/count-votes');
 const urlHash = require('./lib/url-hash');
+const endConfigure = require('./lib/end-configure');
 
 app.locals.polls = {};
 
@@ -56,14 +57,9 @@ io.on('connection', function(socket) {
   socket.on('message', function (channel, message) {
     if (channel === 'voteCast' + message.id) {
       var poll = app.locals.polls[message.id];
+
       if (poll.endTime) {
-        var milTime = moment(poll.endTime).format('x');
-        var timeDone = milTime - moment().format('x');
-          poll.displayTime = moment(poll.endTime).fromNow();
-        setTimeout(function(){
-          poll.status = "closed";
-          io.sockets.emit('pollOver' + poll.id)
-        }, timeDone);
+        endConfigure(poll);
       }
 
       if (moment() >= moment(poll.endTime) || poll.status === "closed") {
